fix: use decimal instead for the finance stuff

+++ Data/ProductRepository.cs
@@ -7,15 +7,14 @@ namespace LegacyOrderService.Data
 {
     public class ProductRepository
     {
-        // TODO: to check this, should we use double? for finance
-        private readonly Dictionary<string, double> _productPrices = new()
+        private readonly Dictionary<string, decimal> _productPrices = new()
         {
-            ["Widget"] = 12.99,
-            ["Gadget"] = 15.49,
-            ["Doohickey"] = 8.75
+            ["Widget"] = 12.99m,
+            ["Gadget"] = 15.49m,
+            ["Doohickey"] = 8.75m
         };
 
-        public double GetPrice(string productName)
+        public decimal GetPrice(string productName)
         {
             // Simulate an expensive lookup
             // TODO: check this and consider to use async instead blocking the main thread
